# specs/001-quant-ops-dashboard/contracts/rest-api.openapi.yaml
openapi: 3.1.0
info:
  title: Kalshi Quant Dashboard API
  version: 0.3.0
  description: |
    Authenticated REST API for historical queries, live or historical convergence,
    admin-managed access policy, feature-flag mutation, alert-rule management,
    role-scoped diagnostics, and export enforcement.
servers:
  - url: https://internal.example.com
    description: Production internal deployment
  - url: http://localhost:3001
    description: Local development
security:
  - cookieAuth: []
paths:
...
  /api/exports/{resource}.csv:
    get:
      summary: Export the current authorized query as CSV
      operationId: exportCsv
      parameters:
        - name: resource
          in: path
          required: true
          schema:
            type: string
            enum: [decisions, trades, skips, alerts, pnl]
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Timezone'
        - $ref: '#/components/parameters/Range'
        - $ref: '#/components/parameters/Strategy'
        - $ref: '#/components/parameters/Symbol'
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Severity'
        - $ref: '#/components/parameters/SkipCategory'
        - $ref: '#/components/parameters/Bucket'
        - $ref: '#/components/parameters/RangeStartUtc'
        - $ref: '#/components/parameters/RangeEndUtc'
        - $ref: '#/components/parameters/Compare'
      responses:
        '200':
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: kqd_session
  parameters:
...
    Sort:
      name: sort
      in: query
      schema:
        type: string
        default: newest
    Search:
      name: search
      in: query
      schema:
        type: string
    Timezone:
      name: timezone
      in: query
      schema:
        type: string
        enum: [utc, local]
        default: utc
...
    Range:
      name: range
      in: query
      schema:
        type: string
        description: Preset range token or custom marker.
    Strategy:
      name: strategy
      in: query
      schema:
        type: string
        description: Comma-separated strategy ids.
    Symbol:
      name: symbol
      in: query
      schema:
        type: string
        description: Comma-separated symbols.
    Market:
      name: market
      in: query
      schema:
        type: string
        description: Comma-separated market tickers.
    Status:
      name: status
      in: query
      schema:
        type: string
        description: Comma-separated lifecycle or alert statuses.
    Severity:
      name: severity
      in: query
      schema:
        type: string
        description: Comma-separated alert severities.
    SkipCategory:
      name: skipCategory
      in: query
      schema:
        type: string
        description: Comma-separated canonical skip categories.
    Bucket:
      name: bucket
      in: query
      schema:
        type: string
        enum: [24h, 7d, 30d, mtd, ytd, all-time, custom]
    RangeStartUtc:
      name: rangeStartUtc
      in: query
      schema:
        type: string
        format: date-time
    RangeEndUtc:
      name: rangeEndUtc
      in: query
      schema:
        type: string
        format: date-time
    Compare:
      name: compare
      in: query
      schema:
        type: string
        description: Comma-separated strategy ids for compare-mode exports.
  responses:
...
    Forbidden:
      description: Authenticated but not authorized for the requested surface
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...